Add get_next_token.c to refactor
