+++ src/lexer/lexer.c
@@ -6,38 +6,13 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/19 14:59:01 by yliu              #+#    #+#             */
-/*   Updated: 2024/08/30 22:50:51 by yliu             ###   ########.fr       */
+/*   Updated: 2024/08/30 22:52:54 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lexer.h"
 #include "lexer_internal.h"
 #include "token.h"
-
-static bool	is_quote_char(char c)
-{
-	return (c == SINGLE_QUOTE || c == DOUBLE_QUOTE);
-}
-
-static t_token_list	*get_next_token(t_lexer *lexer)
-{
-	char	*token_value;
-
-	while (true)
-	{
-		if (*lexer->end == '\0')
-		{
-			token_value = ft_substr(lexer->start, 0, lexer->end - lexer->start);
-			lexer->start = lexer->end;
-			return (construct_token(TOKEN_WORD, token_value));
-		}
-		if (is_quote_char(*lexer->end))
-			process_quote(lexer, *lexer->end);
-		if (ft_isspace(*lexer->end))
-			return (process_blank(lexer));
-		lexer->end++;
-	}
-}
 
 t_token_list	*lexer(const char *string)
 {
