+++ src/lexer/lexer_internal.h
@@ -6,7 +6,7 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/30 22:25:14 by yliu              #+#    #+#             */
-/*   Updated: 2024/08/30 22:48:33 by yliu             ###   ########.fr       */
+/*   Updated: 2024/08/30 22:53:22 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,7 @@ typedef struct s_lexer
 }				t_lexer;
 
 void			construct_lexer(char *string, t_lexer *lexer);
+t_token_list	*get_next_token(t_lexer *lexer);
 t_token_list	*process_blank(t_lexer *lexer);
 void			process_quote(t_lexer *lexer, char starting_quote_char);
